Created GET method for user information

+++ My2Cents.API/My2Cents.DataInfrastructure/ScaffoldModel/UserProfile.cs
@@ -16,12 +16,12 @@ namespace My2Cents.DataInfrastructure
         public string? Email { get; set; }
         public string? SecondaryEmail { get; set; }
         public string? MailingAddress { get; set; }
-        public int? Phone { get; set; }
+        public double? Phone { get; set; }
         public string? City { get; set; }
         public string? State { get; set; }
         public string? Employer { get; set; }
         public string? WorkAddress { get; set; }
-        public int? WorkPhone { get; set; }
+        public double? WorkPhone { get; set; }
 
         public virtual ICollection<Account> Accounts { get; set; }
     }
